Rename strings and tests
